+++ backend/routes/api/session.js
@@ -7,44 +7,7 @@ const { handleValidationErrors } = require("../../utils/validation");
 const router = express.Router();
 
 //=======================================//
-const validateLogin = [
-  check("credential")
-    .exists({ checkFalsy: true })
-    .notEmpty()
-    .withMessage("Please provide a valid email or username."),
-  check("password")
-    .exists({ checkFalsy: true })
-    .withMessage("Please provide a password."),
-  handleValidationErrors
-];
 
-// =============  Log in =================//
-router.post("/login", validateLogin, async (req, res, next) => {
-  const { credential, password } = req.body;
-
-  const user = await User.login({ credential, password });
-
-  if (!user) {
-    const err = new Error("Login failed");
-    err.status = 401;
-    err.title = "Login failed";
-    err.errors = ["The provided credentials were invalid."];
-    return next(err);
-  }
-
-  const token = setTokenCookie(res, user);
-
-  return res.json({
-    user: user.toSafeObject(),
-    token
-  });
-});
-
-// =============== Log out =======================//
-router.delete("/", (_req, res) => {
-  res.clearCookie("token");
-  return res.json({ message: "success" });
-});
 //
 //
 //
