+++ backend/routes/api/users.js
@@ -56,6 +56,45 @@ router.post("/signup", validateSignup, async (req, res, next) => {
     token
   });
 });
+
+// =============  Log in =================//
+const validateLogin = [
+  check("credential")
+    .exists({ checkFalsy: true })
+    .notEmpty()
+    .withMessage("Please provide a valid email or username."),
+  check("password")
+    .exists({ checkFalsy: true })
+    .withMessage("Please provide a password."),
+  handleValidationErrors
+];
+
+router.post("/login", validateLogin, async (req, res, next) => {
+  const { credential, password } = req.body;
+
+  const user = await User.login({ credential, password });
+
+  if (!user) {
+    const err = new Error("Login failed");
+    err.status = 401;
+    err.title = "Login failed";
+    err.errors = ["The provided credentials were invalid."];
+    return next(err);
+  }
+
+  const token = setTokenCookie(res, user);
+
+  return res.json({
+    user: user.toSafeObject(),
+    token
+  });
+});
+
+// =============== Log out =======================//
+router.delete("/", (_req, res) => {
+  res.clearCookie("token");
+  return res.json({ message: "success" });
+});
 //
 
 module.exports = router;
